Item refactoring
